clean up references, and constant variable names
