+++ internal/constants/errors.go
@@ -1,4 +1,4 @@
 package constants
 
-const EMPTY_ENCRYPTION_KEY_ERROR = "ENCRYPTION_KEY not set please, generate an Encryption Key of any length, and set it to continue"
-const EMPTY_SECRET_KEY_ERROR = "SECRET_KEY not set,  generate a Secret Key of any length, and set it to continue"
+const EmptyEncryptionKeyError = "ENCRYPTION_KEY not set please, generate an Encryption Key of any length, and set it to continue"
+const EmptySecretKeyError = "SECRET_KEY not set,  generate a Secret Key of any length, and set it to continue"
