--- internal/constants/icons.go
@@ -1,105 +1,0 @@
-package constants
-
-import "receipt-wrangler/api/internal/structs"
-
-var Icons = []structs.Icon{
-	{Value: "receipt_long", DisplayValue: "Long Receipt"},
-	{Value: "account_balance_wallet", DisplayValue: "Wallet"},
-	{Value: "shopping_cart", DisplayValue: "Shopping Cart"},
-	{Value: "credit_card", DisplayValue: "Credit Card"},
-	{Value: "attach_money", DisplayValue: "Money"},
-	{Value: "local_atm", DisplayValue: "ATM"},
-	{Value: "currency_exchange", DisplayValue: "Currency Exchange"},
-	{Value: "pie_chart", DisplayValue: "Pie Chart"},
-	{Value: "bar_chart", DisplayValue: "Bar Chart"},
-	{Value: "trending_up", DisplayValue: "Trending Up"},
-	{Value: "trending_down", DisplayValue: "Trending Down"},
-	{Value: "compare_arrows", DisplayValue: "Compare"},
-	{Value: "store", DisplayValue: "Store"},
-	{Value: "local_offer", DisplayValue: "Offer"},
-	{Value: "local_mall", DisplayValue: "Mall"},
-	{Value: "loyalty", DisplayValue: "Loyalty"},
-	{Value: "redeem", DisplayValue: "Redeem"},
-	{Value: "group", DisplayValue: "Group"},
-	{Value: "person", DisplayValue: "Person"},
-	{Value: "family_restroom", DisplayValue: "Family"},
-	{Value: "notifications", DisplayValue: "Notifications"},
-	{Value: "calendar_today", DisplayValue: "Calendar"},
-	{Value: "event", DisplayValue: "Event"},
-	{Value: "schedule", DisplayValue: "Schedule"},
-	{Value: "today", DisplayValue: "Today"},
-	{Value: "history", DisplayValue: "History"},
-	{Value: "update", DisplayValue: "Update"},
-	{Value: "sync", DisplayValue: "Sync"},
-	{Value: "cloud_upload", DisplayValue: "Upload"},
-	{Value: "photo_camera", DisplayValue: "Camera"},
-	{Value: "document_scanner", DisplayValue: "Scanner"},
-	{Value: "folder", DisplayValue: "Folder"},
-	{Value: "description", DisplayValue: "Description"},
-	{Value: "list", DisplayValue: "List"},
-	{Value: "checklist", DisplayValue: "Checklist"},
-	{Value: "rule", DisplayValue: "Rule"},
-	{Value: "calculate", DisplayValue: "Calculate"},
-	{Value: "savings", DisplayValue: "Savings"},
-	{Value: "payments", DisplayValue: "Payments"},
-	{Value: "category", DisplayValue: "Category"},
-	{Value: "local_grocery_store", DisplayValue: "Grocery Store"},
-	{Value: "restaurant", DisplayValue: "Restaurant"},
-	{Value: "local_gas_station", DisplayValue: "Gas Station"},
-	{Value: "flight", DisplayValue: "Flight"},
-	{Value: "hotel", DisplayValue: "Hotel"},
-	{Value: "directions_car", DisplayValue: "Car"},
-	{Value: "home", DisplayValue: "Home"},
-	{Value: "work", DisplayValue: "Work"},
-	{Value: "settings", DisplayValue: "Settings"},
-	{Value: "account_balance", DisplayValue: "Account Balance"},
-	{Value: "assignment", DisplayValue: "Assignment"},
-	{Value: "article", DisplayValue: "Article"},
-	{Value: "backup", DisplayValue: "Backup"},
-	{Value: "bookmark", DisplayValue: "Bookmark"},
-	{Value: "bug_report", DisplayValue: "Bug Report"},
-	{Value: "build", DisplayValue: "Build"},
-	{Value: "cached", DisplayValue: "Cached"},
-	{Value: "chat", DisplayValue: "Chat"},
-	{Value: "cloud", DisplayValue: "Cloud"},
-	{Value: "code", DisplayValue: "Code"},
-	{Value: "content_copy", DisplayValue: "Copy"},
-	{Value: "dashboard", DisplayValue: "Dashboard"},
-	{Value: "delete", DisplayValue: "Delete"},
-	{Value: "eco", DisplayValue: "Eco"},
-	{Value: "edit", DisplayValue: "Edit"},
-	{Value: "email", DisplayValue: "Email"},
-	{Value: "euro_symbol", DisplayValue: "Euro"},
-	{Value: "explore", DisplayValue: "Explore"},
-	{Value: "favorite", DisplayValue: "Favorite"},
-	{Value: "flag", DisplayValue: "Flag"},
-	{Value: "flash_on", DisplayValue: "Flash"},
-	{Value: "format_list_bulleted", DisplayValue: "Bullet List"},
-	{Value: "functions", DisplayValue: "Functions"},
-	{Value: "group_add", DisplayValue: "Add Group"},
-	{Value: "help", DisplayValue: "Help"},
-	{Value: "info", DisplayValue: "Info"},
-	{Value: "insert_chart", DisplayValue: "Insert Chart"},
-	{Value: "label", DisplayValue: "Label"},
-	{Value: "language", DisplayValue: "Language"},
-	{Value: "launch", DisplayValue: "Launch"},
-	{Value: "lightbulb", DisplayValue: "Idea"},
-	{Value: "link", DisplayValue: "Link"},
-	{Value: "location_on", DisplayValue: "Location"},
-	{Value: "lock", DisplayValue: "Lock"},
-	{Value: "map", DisplayValue: "Map"},
-	{Value: "menu_book", DisplayValue: "Menu"},
-	{Value: "more_horiz", DisplayValue: "More"},
-	{Value: "note_add", DisplayValue: "Add Note"},
-	{Value: "pending_actions", DisplayValue: "Pending"},
-	{Value: "query_stats", DisplayValue: "Statistics"},
-	{Value: "report", DisplayValue: "Report"},
-	{Value: "search", DisplayValue: "Search"},
-	{Value: "share", DisplayValue: "Share"},
-	{Value: "star", DisplayValue: "Star"},
-	{Value: "summarize", DisplayValue: "Summary"},
-	{Value: "swap_horiz", DisplayValue: "Swap"},
-	{Value: "visibility", DisplayValue: "Visibility"},
-	{Value: "warning", DisplayValue: "Warning"},
-	{Value: "inventory", DisplayValue: "Inventory"},
-}
