--- internal/constants/receipt_status.go
@@ -1,7 +1,0 @@
-package constants
-
-import "receipt-wrangler/api/internal/models"
-
-func ReceiptStatuses() []interface{} {
-	return []interface{}{models.OPEN, models.NEEDS_ATTENTION, models.RESOLVED, models.DRAFT}
-}
