+++ internal/handlers/generic_handler_test.go
@@ -29,7 +29,7 @@ func TestShouldSetContentTypeHeader(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
 			return 0, nil
 		},
@@ -39,8 +39,8 @@ func TestShouldSetContentTypeHeader(t *testing.T) {
 
 	contentType := w.Header().Get("Content-Type")
 
-	if contentType != constants.APPLICATION_JSON {
-		utils.PrintTestError(t, contentType, constants.APPLICATION_JSON)
+	if contentType != constants.ApplicationJson {
+		utils.PrintTestError(t, contentType, constants.ApplicationJson)
 	}
 }
 
@@ -58,7 +58,7 @@ func TestShouldRejectAccessBasedOnGroupId(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.VIEWER,
 		GroupId:      "2",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -87,7 +87,7 @@ func TestShouldRejectAccessBasedOnGroupIdIfGroupDoesNotExist(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.VIEWER,
 		GroupId:      "500",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -116,7 +116,7 @@ func TestShouldRejectAccessBasedOnGroupIdIfGroupIdIsMalformed(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.VIEWER,
 		GroupId:      "bad parse",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -152,7 +152,7 @@ func TestShouldRejectReceiptAccessBasedOnGroup(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.EDITOR,
 		ReceiptId:    "1",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -190,7 +190,7 @@ func TestShouldAcceptReceiptAccessBasedOnGroup(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.OWNER,
 		ReceiptId:    "1",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -219,7 +219,7 @@ func TestShouldAcceptAccessBasedOnGroupId(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.VIEWER,
 		GroupId:      "1",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -263,7 +263,7 @@ func TestShouldAcceptReceiptsAccessBasedOnGroup(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.OWNER,
 		ReceiptIds:   []string{"1", "2"},
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -301,7 +301,7 @@ func TestShouldRejectReceiptAccessBasedOnWrongGroupRole(t *testing.T) {
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.OWNER,
 		ReceiptId:    "1",
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
@@ -345,7 +345,7 @@ func TestShouldRejectReceiptAccessBasedOnWrongGroupRoleForMultipleReceipts(t *te
 	handler := structs.Handler{
 		Writer:       w,
 		Request:      r,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		GroupRole:    models.OWNER,
 		ReceiptIds:   []string{"1", "2"},
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
