+++ internal/handlers/imports.go
@@ -15,7 +15,7 @@ func ImportConfigJson(w http.ResponseWriter, r *http.Request) {
 		Writer:       w,
 		Request:      r,
 		UserRole:     models.ADMIN,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
 			command := commands.ConfigImportCommand{}
 			err := command.LoadDataFromRequest(w, r)
