+++ internal/handlers/system_task.go
@@ -16,7 +16,7 @@ func GetSystemTasks(w http.ResponseWriter, r *http.Request) {
 		Writer:       w,
 		Request:      r,
 		UserRole:     models.ADMIN,
-		ResponseType: constants.APPLICATION_JSON,
+		ResponseType: constants.ApplicationJson,
 		HandlerFunction: func(w http.ResponseWriter, r *http.Request) (int, error) {
 			command := commands.GetSystemTaskCommand{}
 			err := command.LoadDataFromRequest(w, r)
